+++ rust/testsuite/src/main.rs
@@ -15,8 +15,7 @@ const CUSTOM_PARAMS_JSON_FLAG: &str = "custom-params-json";
 const KURTOSIS_API_SOCKET_FLAG: &str  = "kurtosis-api-socket";
 const LOG_LEVEL_FLAG: &str = "log-level";
 
-#[tokio::main]
-async fn main() -> Result<()> {
+fn main() -> Result<()> {
     let matches = App::new("My Super Program")
         .arg(Arg::new(CUSTOM_PARAMS_JSON_FLAG)
             .long(CUSTOM_PARAMS_JSON_FLAG)